refactor: split subcommand
