+++ src/cmd/mod.rs
@@ -1,0 +1,2 @@
+pub mod add;
+pub mod list;
