+++ src/main.rs
@@ -1,7 +1,8 @@
-use rbm::bookmarks::{add_bookmark, list_bookmarks};
 use rbm::cli::{Action::*, CommandLineArgs};
 use rbm::config::load_config;
 use structopt::StructOpt;
+
+mod cmd;
 
 fn main() -> anyhow::Result<()> {
     let CommandLineArgs { action } = CommandLineArgs::from_args();
@@ -9,8 +10,8 @@ fn main() -> anyhow::Result<()> {
     let config = load_config().expect("Failed to load the config file");
 
     match action {
-        Add => add_bookmark(config.path),
-        List => list_bookmarks(config.path),
+        Add => cmd::add::add_bookmark(config.path),
+        List => cmd::list::list_bookmarks(config.path),
     }?;
     Ok(())
 }
